Move defaultContext and minimalContext from won.js to vocab.js
