+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/connection-agreement-details.jsx
@@ -5,7 +5,7 @@ import React, { useState, useEffect, useRef } from "react";
 import * as N3 from "n3";
 import { usePrevious } from "../cstm-react-utils.js";
 import PropTypes from "prop-types";
-import won from "../service/won";
+import vocab from "../service/vocab";
 import * as connectionUtils from "../redux/utils/connection-utils.js";
 
 import "../../style/_connection-agreement-details.scss";
@@ -19,7 +19,7 @@ export default function WonConnectionAgreementDetails({ connection }) {
 
   const writer = new N3.Writer({
     format: "application/trig",
-    prefixes: won.minimalContext,
+    prefixes: vocab.minimalContext,
   });
 
   if (
