+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-lunch.js
@@ -8,7 +8,6 @@ import {
   sparqlQuery,
 } from "../../app/sparql-builder-utils.js";
 import { getIn } from "../../app/utils.js";
-import won from "../../app/service/won.js";
 import vocab from "../../app/service/vocab.js";
 import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
 import ico36_detail_datetime from "../../images/won-icons/ico36_detail_datetime.svg";
@@ -56,9 +55,9 @@ export const lunchPlan = {
       bindScoreAs: "?location_geoScore",
       pathToGeoCoords: "s:location/s:geo",
       prefixesInPath: {
-        s: won.defaultContext["s"],
-        won: won.defaultContext["won"],
-        con: won.defaultContext["con"],
+        s: vocab.defaultContext["s"],
+        won: vocab.defaultContext["won"],
+        con: vocab.defaultContext["con"],
       },
       geoCoordinates: getIn(draft, ["content", "location"]),
     });
@@ -72,14 +71,14 @@ export const lunchPlan = {
 
     return sparqlQuery({
       prefixes: {
-        won: won.defaultContext["won"],
-        rdf: won.defaultContext["rdf"],
-        buddy: won.defaultContext["buddy"],
-        hold: won.defaultContext["hold"],
-        s: won.defaultContext["s"],
-        match: won.defaultContext["match"],
-        demo: won.defaultContext["demo"],
-        "wx-persona": won.defaultContext["wx-persona"],
+        won: vocab.defaultContext["won"],
+        rdf: vocab.defaultContext["rdf"],
+        buddy: vocab.defaultContext["buddy"],
+        hold: vocab.defaultContext["hold"],
+        s: vocab.defaultContext["s"],
+        match: vocab.defaultContext["match"],
+        demo: vocab.defaultContext["demo"],
+        "wx-persona": vocab.defaultContext["wx-persona"],
       },
       distinct: true,
       variables: [resultName, "?score"],
@@ -138,9 +137,9 @@ export const lunchInterest = {
       bindScoreAs: "?location_geoScore",
       pathToGeoCoords: "s:location/s:geo",
       prefixesInPath: {
-        s: won.defaultContext["s"],
-        won: won.defaultContext["won"],
-        con: won.defaultContext["con"],
+        s: vocab.defaultContext["s"],
+        won: vocab.defaultContext["won"],
+        con: vocab.defaultContext["con"],
       },
       geoCoordinates: getIn(draft, ["content", "location"]),
     });
@@ -154,11 +153,11 @@ export const lunchInterest = {
 
     return sparqlQuery({
       prefixes: {
-        won: won.defaultContext["won"],
-        rdf: won.defaultContext["rdf"],
-        buddy: won.defaultContext["buddy"],
-        hold: won.defaultContext["hold"],
-        s: won.defaultContext["s"],
+        won: vocab.defaultContext["won"],
+        rdf: vocab.defaultContext["rdf"],
+        buddy: vocab.defaultContext["buddy"],
+        hold: vocab.defaultContext["hold"],
+        s: vocab.defaultContext["s"],
       },
       distinct: true,
       variables: [resultName, "?score"],
