+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/uc-rehearsal-room-offer.js
@@ -11,7 +11,6 @@ import {
   realEstateNumberOfRoomsDetail,
   realEstateFeaturesDetail,
 } from "../details/real-estate.js";
-import won from "../../app/service/won.js";
 import vocab from "../../app/service/vocab.js";
 import { perHourRentDetail } from "../details/musician.js";
 import * as jsonLdUtils from "../../app/service/jsonld-utils.js";
@@ -90,14 +89,14 @@ export const rehearsalRoomOffer = {
         {
           // to select is-branch
           prefixes: {
-            won: won.defaultContext["won"],
-            rdf: won.defaultContext["rdf"],
-            sh: won.defaultContext["sh"], //needed for the filterNumericProperty calls
-            s: won.defaultContext["s"],
+            won: vocab.defaultContext["won"],
+            rdf: vocab.defaultContext["rdf"],
+            sh: vocab.defaultContext["sh"], //needed for the filterNumericProperty calls
+            s: vocab.defaultContext["s"],
             geo: "http://www.bigdata.com/rdf/geospatial#",
             xsd: "http://www.w3.org/2001/XMLSchema#",
-            match: won.defaultContext["match"],
-            demo: won.defaultContext["demo"],
+            match: vocab.defaultContext["match"],
+            demo: vocab.defaultContext["demo"],
           },
           operations: [
             `${resultName} a won:Atom.`,
@@ -130,11 +129,11 @@ export const rehearsalRoomOffer = {
         {
           // to select is-branch
           prefixes: {
-            won: won.defaultContext["won"],
-            rdf: won.defaultContext["rdf"],
-            sh: won.defaultContext["sh"], //needed for the filterNumericProperty calls
-            match: won.defaultContext["match"],
-            demo: won.defaultContext["demo"],
+            won: vocab.defaultContext["won"],
+            rdf: vocab.defaultContext["rdf"],
+            sh: vocab.defaultContext["sh"], //needed for the filterNumericProperty calls
+            match: vocab.defaultContext["match"],
+            demo: vocab.defaultContext["demo"],
           },
           operations: [
             `${resultName} a won:Atom.`,
